Hud in jsbsim
I sleuthed out the problem with the HUD and discovered the problem.  It had not been included in the JSBsim C-17-set.xml file.  I haven't figured out how to turn it off, but it is now visible.  Using the CTRL-Rt Click, you can adjust its position on the glass screen.  Importantly, it is always oriented to the horizon (I think).