Add Loadmaster Station view to C-17 cargo hold.
This commit adds a new view station in the C-17 for the Loadmaster.
The camera is located at the front of the hold on the left (port) side
of the station.  The default view looks toward the hold doors down the
left wall past any vehicle in the hold.  Additional work will need to
be done to ensure the correct pathway to the View/View Options dialog
in the primary menu ribbon.