Loadmaster view (#66)
* Adjusting "Loadmaster" view and field of view.

Test 1: Adjusted z-offset from 25.7 to 10.0 (using zoom in-game).
Adjusted default field of view from 45.3 degrees to 10.3 degrees.

* Add Loadmaster Station view to C-17 cargo hold.

This commit adds a new view station in the C-17 for the Loadmaster.
The camera is located at the front of the hold on the left (port) side
of the station.  The default view looks toward the hold doors down the
left wall past any vehicle in the hold.  Additional work will need to
be done to ensure the correct pathway to the View/View Options dialog
in the primary menu ribbon.

* Renumbering the views

Loadmaster view and Stations are now together

* Loadmaster station is inside looking outside

* Some clean up of view code

* Revision to Loadmaster View.

A few changes in POV values to improve overall view of hold from the
base of the open ramp.

* Fix out-of-alignment LoadmasterStation View.

Changes in the offsets misligned the POV for the Loadmaster Station.
The offsets were realigned in the cargo hold looking rearward from the
front of the hold on the left (port) side of the aircraft where the Loadmaster's station is located.

* Bring station master inside again

* Fine-tuning to the Loadmaster and Loadmaster station views.

More fine-tuning of the Loadmaster station to get the camera and POV
positioned so the Loadmaster can look past the vehicle down the
left (port) side of the aircraft and see the hold bay doors (and out
the doors if they are open.

---------

Co-authored-by: Israel Hernandez <israel.a.hernandez@gmail.com>