2Change C-17 tail camera view for a better perspective
I moved the tail camera so that it is situated on the forward edge of
the vertical stabilizer. Previously the camera was located floating in
the air in front of the tail.  It is now positioned at the surface and
in front of the red beacon light.  Rotating the camera now provides a
good view of nearly 360 degrees around the tail of the C-17